Proposed fix for Issue 11773.
This change updates readyness check in ready.go to correctly
account for statefulsets that are uttilizing a partitioned upgrade.
These statefulsets only upgrade a subset of the managed pods with each call
to helm upgrade. This causes the upgrade to legitimately hit the condition where
sts.status.CurrentRevision != sts.Status.UpdateRevision which causes helm to mark
the upgrade as failed when infact it is succesful.

This change fixes that behavior to only check when parition is unspecified or 0.